feat: ProjectStore handle+channel, PTY shutdown, worktree to roux-core (#27, #28, #29)
Three improvements in one commit:

1. Convert ProjectStore to handle+channel (#27):
   Same pattern as SessionStore — message-passing service with async
   CRUD, periodic persist via spawn_blocking, explicit shutdown.

2. Add PtyManager.shutdown_all() (#28):
   Kills all active PTY sessions on app exit. Called from quit_app
   before service shutdowns.

3. Move worktree operations to roux-core (#29):
   create/remove/list_worktrees + WorktreeError + all helpers and tests
   now live in roux-core. src-tauri re-exports. CLI can reuse directly.

Closes #27, closes #28, closes #29
