+++ crates/roux-core/Cargo.toml
@@ -8,3 +8,5 @@ description = "Shared types and models for Roux"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 specta = { version = "=2.0.0-rc.24", features = ["derive", "serde_json"] }
+thiserror = "2"
+dirs = "6"
