+++ crates/roux-core/src/lib.rs
@@ -5,5 +5,7 @@
 //! Tauri, so it can be used by the CLI, tests, and future tooling.
 
 pub mod models;
+pub mod worktree;
 
 pub use models::*;
+pub use worktree::{create_worktree, remove_worktree, list_worktrees, WorktreeError};
